+++ src/frontend/servers/y-provider/nodemon.json
@@ -1,5 +1,5 @@
 {
   "watch": ["src"],
   "ext": "ts",
-  "exec": "yarn build"
+  "exec": "yarn build && yarn start"
 }
