+++ src/frontend/servers/y-provider/src/start-server.ts
@@ -1,3 +1,3 @@
-import { initServer } from './server';
+import { initServer } from './servers/appServer';
 
 initServer();
