+++ src/frontend/servers/y-provider/src/utils.ts
@@ -1,9 +1,9 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { COLLABORATION_LOGGING } from './env';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function logger(...args: any[]) {
   if (COLLABORATION_LOGGING === 'true') {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
     console.log(...args);
   }
 }
